@@ -1,9 +1,15 @@
-<div class='flex self-center items-center justify-center text-grayedOut_dark gap-2 md:gap-4'>
-	<a href='/impressum'>Impressum</a>
+<div class="flex items-center justify-center gap-2 self-center text-grayedOut_dark md:gap-4">
+	<a href="/impressum">Impressum</a>
 	<p>•</p>
-	<a href='/datenschutz'>Datenschutz</a>
+	<a href="/datenschutz">Datenschutz</a>
 	<p>•</p>
-	<a href='https://cubyx.eu/status' target='_blank'>Statuspage</a>
+	<a href="https://cubyx.eu/status" target="_blank">Statuspage</a>
 	<p>•</p>
-	<a href='/kontakt'>Kontakt</a>
+	<a href="/kontakt">Kontakt</a>
 </div>
+
+<style>
+	a {
+		@apply text-grayedOut_light;
+	}
+</style>
